added test timeouts to the ViennaCL's jwein2 tests
it seems that a test (sparse-opencl) on optimized-clang did not finish over night,
therefore only one test got submitted to CDash (clang-debug) as this one was executed before the optimized one.
in an attempt to avoid such a 'deadlock' I have added a CTEST_TEST_TIMEOUT parameter and set it to 500 seconds.
let's see whether this does the trick.

however, no clue why the test 'deadlocked'. it finished just fine when re-executed .