deactivated clang-optimized test on jwein2 as it deadlocks
so, here is the puzzle: is the test (or the entire script) executed manually, all is fine.
however, if it is executed by the cronjob a specific ViennaCL test (sparse-test-opencl) 'deadlocks' (i.e. it runs forever..)
the debug clang test runs through without a problem..

now let's see if the other compilers have the same problem